requirements.txt
# Requirements for hw2.py
torch
torchvision
requests
Pillow